SE-2690 Python 3 Compatibility for Juniper Upgrade (#3)
Updated the Vimeo XBlock to work with the Juniper Release. In addition, maintained support for Python 2.7 for Ironwood release. 

Adding testing to the library through tox. Unfortunately, testing only supports python 3.5 and 3.8. However, manual testing was done for python 2.7.

Progress View through Commit Messages:

* Switched UrlParse to UrlLib.Parse for Python 3

Added check for python version through `sys.version_info`. If it is python 3, then use urllib.parse instead of urlparse.

In python 3, the urlparse library was renamed to urllib.parse.

* Unified Fragment Creation in Method and Replaced Unicode with Resource String Method

Previously, each fragment was being created with 2 separate codes. This could be avoided by using a single method to create the fragment. In addition, the and use of unicode() isn't necessary. The string is already being utf8 decoded thanks to the resource_string method. Also, load_resource and resource_string were doing the same work, which was another reason to combine the fragment creation.

* Updated Setup to v0.5 and Added edX Release Tags

* Defined the Error Messages for Getting Embeded Code in Object

This will make it easier to change later on, without affecting the test cases, since the test cases will reference those error messagses from inside the object

* Added Tox Automated Testing with Unit Tests for Student and Studio View

* Added Back load_resource for Python 2.7 HTML Content not Encoded with UTF-8

* Updated README with Testing Instructions and Installing XBlock on DevStack

* Added Extra Step to Run Installation on DevStack for LMS

Previously, it was only for Studio, but the installation should be done on both the LMS and the Studio

* Added Empty Line at the End of the Tox.ini File and Test File

* Moved XBlock Vimeo Test Instance Creator to Utils File

* Changed Mock Create Fragment from an Assert Return Value to an Assert Called Once Call for Studio and Student View

* Added Fragment Return Value to Mock Create Fragment

This makes it less confusing as to what the result should be at the end. The create_fragment method returns a fragment, therefore, its mock should too

* Fixed Extra Space for Test Requirements File Dependency